pyproject.toml
[project]
name = "hf-pe-industry-predictability"
version = "0.1.0"
description = "Historical US sector PE ratios and returns pipeline"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "wrds>=3.2.0",
    "polars>=1.0.0",
    "pyarrow>=16.0.0",
    "pandas>=2.2.0",
    "numpy>=2.0.0",
    "rich>=13.0.0",
    "typer[all]>=0.12.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]
